+++ src/app/patterns/bridge/bridge.module.ts
