+++ src/app/patterns/bridge/view.component.ts
