+++ src/app/patterns/bridge/widget-wrapper/widget-wrapper.component.scss
