+++ src/app/patterns/bridge/widgets/velocity-widget/velocity-widget.component.html
