+++ src/app/patterns/bridge/widgets/weather-widget/weather-widget.component.html
