+++ src/app/patterns/bridge/widgets/widget.interface.ts
