+++ src/app/patterns/bridge/widgets/widget.token.ts
