+++ src/app/patterns/composition/composition.component.spec.ts
