+++ src/app/patterns/composition/composition.component.ts
