+++ src/app/patterns/composition/composition.module.ts
