+++ src/app/patterns/standalone/standalone.component.spec.ts
