+++ src/app/patterns/standalone/standalone.component.ts
