+++ tsconfig.json
@@ -1,4 +1,4 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
+/** To learn more about this file see: https://angular.io/config/tsconfig. **/
 {
   "compileOnSave": false,
   "compilerOptions": {
